linters and restructuring
