+++ main.go
@@ -27,12 +27,11 @@ func main() {
 
 	bin, err := utils.NewEvmosTestingBinary()
 	if err != nil {
-		log.Fatalf("Error creating binary: %v", err)
+		log.Fatalf("error creating binary: %v", err)
 	}
 
-	err = bin.GetAccounts()
-	if err != nil {
-		log.Fatalf("Error getting accounts: %v", err)
+	if err = bin.GetAccounts(); err != nil {
+		log.Fatalf("error getting accounts: %v", err)
 	}
 
 	// TODO: use with Cobra CLI
@@ -40,34 +39,32 @@ func main() {
 	case "vote":
 		proposalID, err := getProposalIDFromInput(bin, os.Args)
 		if err != nil {
-			log.Fatalf("Error getting proposal ID: %v", err)
+			log.Fatalf("error getting proposal ID: %v", err)
 		}
 
-		err = gov.SubmitAllVotesForProposal(bin, proposalID)
-		if err != nil {
-			log.Fatalf("Error submitting votes for proposal %d: %v", proposalID, err)
+		if err = gov.SubmitAllVotesForProposal(bin, proposalID); err != nil {
+			log.Fatalf("error submitting votes for proposal %d: %v", proposalID, err)
 		}
 
 	case "deposit":
 		proposalID, err := getProposalIDFromInput(bin, os.Args)
 		if err != nil {
-			log.Fatalf("Error getting proposal ID: %v", err)
+			log.Fatalf("error getting proposal ID: %v", err)
 		}
 
 		// TODO: replace fixed amount with min deposit from chain params
 		if _, err = gov.DepositForProposal(bin, proposalID, bin.Accounts[0].Name, 1e9); err != nil {
-			log.Fatalf("Error depositing for proposal %d: %v", proposalID, err)
+			log.Fatalf("error depositing for proposal %d: %v", proposalID, err)
 		}
 
 	default:
 		targetVersion := os.Args[1]
 		if matched, _ := regexp.MatchString(`v\d+\.\d+\.\d(-rc\d+)?`, targetVersion); !matched {
-			log.Fatalf("Invalid target version: %s. Please use the format vX.Y.Z(-rc*).\n", targetVersion)
+			log.Fatalf("invalid target version: %s; please use the format vX.Y.Z(-rc*).\n", targetVersion)
 		}
 
-		err := upgradeLocalNode(bin, targetVersion)
-		if err != nil {
-			log.Fatalf("Error upgrading local node: %v", err)
+		if err := upgradeLocalNode(bin, targetVersion); err != nil {
+			log.Fatalf("error upgrading local node: %v", err)
 		}
 	}
 }
@@ -83,15 +80,15 @@ func getProposalIDFromInput(bin *utils.Binary, args []string) (int, error) {
 	case 2:
 		proposalID, err = gov.QueryLatestProposalID(bin)
 		if err != nil {
-			return 0, errors.Wrap(err, "Error querying latest proposal ID")
+			return 0, errors.Wrap(err, "error querying latest proposal ID")
 		}
 	case 3:
 		proposalID, err = strconv.Atoi(args[2])
 		if err != nil {
-			return 0, errors.Wrapf(err, "Error converting proposal ID %s to integer", args[2])
+			return 0, errors.Wrapf(err, "error converting proposal ID %s to integer", args[2])
 		}
 	default:
-		return 0, errors.New("Invalid number of arguments")
+		return 0, errors.New("invalid number of arguments")
 	}
 
 	return proposalID, nil
@@ -102,7 +99,7 @@ func getProposalIDFromInput(bin *utils.Binary, args []string) (int, error) {
 func upgradeLocalNode(bin *utils.Binary, targetVersion string) error {
 	currentHeight, err := utils.GetCurrentHeight(bin)
 	if err != nil {
-		return errors.Wrap(err, "Error getting current height")
+		return errors.Wrap(err, "error getting current height")
 	}
 
 	upgradeHeight := currentHeight + deltaHeight
@@ -111,14 +108,13 @@ func upgradeLocalNode(bin *utils.Binary, targetVersion string) error {
 
 	proposalID, err := gov.SubmitUpgradeProposal(bin, targetVersion, upgradeHeight)
 	if err != nil {
-		return errors.Wrap(err, "Error executing upgrade proposal")
+		return errors.Wrap(err, "error executing upgrade proposal")
 	}
 
 	log.Printf("Scheduled upgrade to %s at height %d.\n", targetVersion, upgradeHeight)
 
-	err = gov.SubmitAllVotesForProposal(bin, proposalID)
-	if err != nil {
-		return errors.Wrapf(err, "Error submitting votes for proposal %d", proposalID)
+	if err = gov.SubmitAllVotesForProposal(bin, proposalID); err != nil {
+		return errors.Wrapf(err, "error submitting votes for proposal %d", proposalID)
 	}
 
 	return nil
